+++ src/services/unitService.js
@@ -1,5 +1,9 @@
 import APIFeatures from "@/lib/apiFeatues";
 import { catchAsync } from "@/lib/catchAsync";
+import { Buy } from "@/models/Buy";
+import { DepotItems } from "@/models/depotItems";
+import { Items } from "@/models/items";
+import { Sale } from "@/models/Sale";
 import { Unit } from "@/models/unit";
 //////////////////////////// CREATE ///////////////////////////////////////
 export const createUnit = catchAsync(async (data) => {
@@ -29,6 +33,23 @@ export const updateUnit = catchAsync(async (data) => {
 
 ////////////////////////////DELETE ////////////////////////////
 export const deleteUnit = catchAsync(async (_id) => {
+  const buy = await Buy.findOne({
+    "items.unit": _id,
+  });
+  const sale = await Items.findOne({
+    unit: _id,
+  });
+  const item = await Sale.findOne({
+    "items.unit": _id,
+  });
+  const depotItems = await DepotItems.findOne({
+    unit: _id,
+  });
+  if (buy || sale || item || depotItems)
+    return {
+      message:
+        "نمی توانین این عنوان را پاک کنید چون با این عنوان محصول در خرید فروش یا در انبار ثبت شده",
+    };
   const result = await Unit.findByIdAndDelete(_id);
   return result;
 });
